fix: 修复代码编译失败的bug[2]

+++ src/router.cc
@@ -33,13 +33,13 @@ void Router::route() {
                 if (ip_datagram.header.ttl-- <= 0) continue;/*不转发ttl过期的ip数据报*/
 
                 // 找到next hop & interface id
-                uint8_t max_prefix_length = -1;
+                int8_t max_prefix_length = -1;
                 route_t ans_r{};
                 for (route_t const &r: routing_table_) {
                     uint8_t const offset = static_cast<uint8_t>(32) - r.prefix_length;
                     if ((r.route_prefix >> offset) == (ip_datagram.header.dst >> offset)) {/*前缀匹配*/
                         if (max_prefix_length < r.prefix_length) {/*取最长前缀*/
-                            max_prefix_length = r.prefix_length;
+                            max_prefix_length = static_cast<int8_t>(r.prefix_length);
                             ans_r = r;
                         }
                     }
